windows_mt5_server/requirements.txt
# Python packages
MetaTrader5>=5.0.45
flask>=2.0.1
pandas>=1.3.0
pytest>=6.2.5
requests>=2.26.0
python-dotenv>=0.19.0